requirements.txt
RouterOS-api
yeelight
requests
apistar
click
linode-cli
broadlink==0.12.0
pycrypto
tldextract
pywin32;platform_system=="Windows"
PyXiaomiGateway
boto3
